+++ test/unit/fuzz-tools.test.js
@@ -116,6 +116,7 @@ import * as pulmV114 from '../../lib/pulm-v114.js';
 import * as pulmnodV115 from '../../lib/pulmnod-v115.js';
 import * as neuroV117 from '../../lib/neuro-v117.js';
 import * as neuroV118 from '../../lib/neuro-v118.js';
+import * as neuroV119 from '../../lib/neuro-v119.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -176,6 +177,7 @@ const MODULES = {
   'pulmnod-v115.js': pulmnodV115,
   'neuro-v117.js': neuroV117,
   'neuro-v118.js': neuroV118,
+  'neuro-v119.js': neuroV119,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];